+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasetop.py
@@ -83,5 +83,5 @@ class CreateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-21T11:42:37.749Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-21T12:58:34.536Z")
     r"""The date and time the resource was last updated"""