+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -125,7 +125,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01JVSAEB7SQJESFVBASQDWHR37"
+        "01JVSESDPRJX6320YY0QT98EVE"
     )
     r"""The id of the resource"""