+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-05-21T11:42:37.749Z")
+    updated: Optional[datetime] = parse_datetime("2025-05-21T12:58:34.536Z")
     r"""The date and time the resource was last updated"""